Improve code quality and ignore config file if OneSixLauncher is detected
