+++ src/main/resources/lang/fb.properties
@@ -11,6 +11,7 @@ server-pinging=Pinging {0}:{1} with protocol of MC-{2}
 packet-could-not-be-received=Could not receive packet! Shutting down...
 bot-could-not-be-started=An unexpected error occured: {0}\nThe bot will be stopped...
 bot-automatic-reconnect=FishingBot will reconnect in {0} seconds...
+auth-change-onesix=Changing authentication to use OneSix
 auth-using-onesix=Trying to authenticate using OneSix arguments...
 auth-using-access-token=Trying to authenticate using access token...
 auth-invalid-access-token=The access token is invalid: {0} - Trying to refresh...
