+++ src/main/resources/lang/fb_en.properties
@@ -14,6 +14,8 @@ packet-could-not-be-received=Could not receive packet! Shutting down...
 bot-could-not-be-started=An unexpected error occured: {0}\n\
 The bot will be stopped...
 bot-automatic-reconnect=FishingBot will reconnect in {0} seconds...
+auth-change-onesix=Changing authentication to use OneSix
+auth-using-onesix=Trying to authenticate using OneSix arguments...
 auth-using-access-token=Trying to authenticate using access token...
 auth-invalid-access-token=The access token is invalid: {0} - Trying to refresh...
 auth-successful=Authentication successful!
